CaregivingLM/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchainhub>=0.0.1
langchain-core>=0.1.0
langchain-chroma>=0.0.1
openai>=1.0.0
chromadb>=0.4.0
pypdf>=3.0.0
tiktoken>=0.5.0
numpy>=1.24.0
tqdm>=4.65.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0
python-dotenv>=1.0.0
paperscraper>=0.1.0